feat: Update bimbingan_tugas_akhir method in PembimbingMahasiswaController
Refactor the bimbingan_tugas_akhir method in PembimbingMahasiswaController to allow for selecting a specific semester. If a semester is provided in the request, use that semester for retrieving the data. Otherwise, use the active semester. This change improves the flexibility of the method and allows for better customization of the displayed data.