Refactor KHSController and index.blade.php
The code changes in this commit refactor the KHSController and index.blade.php files. It includes commenting out the 'where' condition for 'id_semester' in the '$nilai' and '$konversi' queries in KHSController.php, and commenting out the 'div' element with class 'col-md-4' containing the 'semester' select input in index.blade.php. These changes are made to temporarily disable the filtering of data by semester in the KHS (Academic Transcript) feature.